CryptoYaTicker: Make time long
The time has to be a long because the API returns a timestamp.

+++ src/main/java/bisq/price/util/cryptoya/CryptoYaTicker.java
@@ -32,6 +32,6 @@ public class CryptoYaTicker {
 
     private Double totalBid;
 
-    private Integer time;
+    private long time;
 
 }
